Implement long polling in web interface
Hold /get request for up to 10 seconds waiting for a change rather than
return immediately.  Also, change the Clocks to report to the
XmlScoreBoardServlet only whole second changes.  Limits updates sent
across the network to 1 per second (if clocks are synced) or 1 per
second per running clock (if clocks are not synced)